+++ forms-flow-api/src/formsflow_api/resources/formio.py
@@ -24,7 +24,6 @@ from formsflow_api_utils.utils.startup import (
 )
 from formsflow_api_utils.utils.user_context import UserContext, user_context
 
-
 API = Namespace("Formio", description="formio")