+++ forms-flow-web/src/components/Modeler/index.js
@@ -14,7 +14,7 @@ import {
   createNewProcess,
   extractDataFromDiagram,
 } from "./helpers/helper";
-import "./Modeller.scss";
+import "./Modeler.scss";
 
 import {
   fetchAllBpmProcesses,
@@ -33,14 +33,14 @@ export default React.memo(() => {
   const dmn = useSelector((state) => state.process.dmnProcessList);
   const [processList, setProcessList] = useState(listProcess(process));
   const workflow = useSelector((state) => state.process.workflowAssociated);
-  const [showModeller, setShowModeller] = useState(false);
+  const [showModeler, setShowModeler] = useState(false);
   const [isBpmnModel, setIsBpmnModel] = useState(true);
   const [isNewDiagram, setIsNewDiagram] = useState(false);
   const tenantKey = useSelector((state) => state.tenants?.tenantId);
 
   useEffect(() => {
     setIsNewDiagram(false);
-    setShowModeller(false);
+    setShowModeler(false);
     dispatch(setWorkflowAssociation(null));
     dispatch(fetchAllBpmProcesses(tenantKey));
   }, []);
@@ -58,26 +58,11 @@ export default React.memo(() => {
   }, [isBpmnModel]);
 
   const handleListChange = (item) => {
-
-    // Check to see if the selected process/definition is already rendered in the canvas
-    // Resolves the false 'ID must be unique' error in the properties panel
-    const object = workflow?.deployedDefinitions;
-    let found = false;
-    if (object){
-      for (const value of Object.values(object)) { 
-        if (value?.key == item.value){
-          found = true;
-        }
-      }
-    }
-
-    if (!found){
-      setIsNewDiagram(false);
-      setShowModeller(true);
-      dispatch(setWorkflowAssociation(item));
-      dispatch(setProcessDiagramXML(null));
-      showChosenFileName(item);
-    }
+    setIsNewDiagram(false);
+    setShowModeler(true);
+    dispatch(setWorkflowAssociation(item));
+    dispatch(setProcessDiagramXML(null));
+    showChosenFileName(item);
   };
 
   const showChosenFileName = (item) => {
@@ -124,11 +109,11 @@ export default React.memo(() => {
     };
     dispatch(setWorkflowAssociation(newWorkflow));
     dispatch(setProcessDiagramXML(newWorkflow.xml));
-    setShowModeller(true);
+    setShowModeler(true);
   };
 
   const handleChangeFile = (file) => {
-    setShowModeller(false);
+    setShowModeler(false);
     setIsNewDiagram(true);
     let fileData = new FileReader();
     try {
@@ -146,13 +131,13 @@ export default React.memo(() => {
     const newProcess = isBpmnModel ? createNewProcess() : createNewDecision();
     dispatch(setWorkflowAssociation(newProcess.defaultWorkflow));
     dispatch(setProcessDiagramXML(newProcess.defaultWorkflow.xml));
-    setShowModeller(true);
+    setShowModeler(true);
     document.getElementById("inputWorkflow").value = null;
   };
 
   const handleToggle = () => {
     setIsNewDiagram(false);
-    setShowModeller(false);
+    setShowModeler(false);
     dispatch(setWorkflowAssociation(null));
     setIsBpmnModel((toggle) => !toggle);
     document.getElementById("inputWorkflow").value = null;
@@ -162,11 +147,18 @@ export default React.memo(() => {
     console.log(message, err);
     document.getElementById("inputWorkflow").value = null;
     dispatch(setWorkflowAssociation(null));
-    setShowModeller(false);
+    setShowModeler(false);
   };
 
   const handleHelp = () => {
     window.open("https://camunda.com/bpmn/");
+  };
+
+  const customDropdownStyles = {
+    menuList: base => ({
+      ...base,
+      maxHeight: "170px",
+    })
   };
 
   return (
@@ -186,7 +178,6 @@ export default React.memo(() => {
           </div>
         </div>
       </div>
-
       <Grid
         container
         direction="row"
@@ -209,16 +200,16 @@ export default React.memo(() => {
                 </span>
                 <div className="select-style">
                   <Select
-                    placeholder={t("Select...")}
-                    dropdownHeight={"135px"}
+                    placeholder={t("Select ...")}
                     options={processList}
                     onChange={handleListChange}
                     value={
                       processList.length && workflow?.value ? workflow : ""
                     }
+                    styles={customDropdownStyles}
                   />
                 </div>
-                <div className="mt-2">
+                <div className="mt-2 toggle-bpm">
                   <label className="switch">
                     <input
                       type="checkbox"
@@ -234,23 +225,21 @@ export default React.memo(() => {
                   </label>
                 </div>
               </Grid>
-
-              <div className="create-import-container">
+              <div className="mt-2">
                 <span className="fontsize-16">
                   {t(
                     "Or create new workflow or import a workflow from a local directory."
                   )}
                 </span>
 
-                <div className="create-import-btns-container">
+                <div className="create-import-btns-container mt-2 mb-4">
                   <Button
-                    className="btn-create-new"
+                    className="btn-create-new mr-3"
                     onClick={() => handleCreateNew()}
                   >
                     {t("Create New")}
                   </Button>
 
-                  <span className="fontsize-16 or-txt">{t(" ")}</span>
                   <input
                     id="inputWorkflow"
                     type="file"
@@ -260,18 +249,18 @@ export default React.memo(() => {
                 </div>
               </div>
 
-              {processList.length && workflow?.value && showModeller ? (
+              {processList.length && workflow?.value && showModeler ? (
                 <div>
                   {isBpmnModel ? (
                     <BpmnEditor
-                      setShowModeller={setShowModeller}
+                      setShowModeler={setShowModeler}
                       processKey={workflow?.value}
                       tenant={workflow?.tenant}
                       isNewDiagram={isNewDiagram}
                     />
                   ) : (
                     <DmnEditor
-                      setShowModeller={setShowModeller}
+                      setShowModeler={setShowModeler}
                       processKey={workflow?.value}
                       tenant={workflow?.tenant}
                       isNewDiagram={isNewDiagram}